+++ components/molecules/search/search.types.ts
@@ -1,5 +1,5 @@
 import { DropDownPickerProps } from 'react-native-dropdown-picker';
 
-export type SearchProps = DropDownPickerProps;
+export type SearchProps = DropDownPickerProps<any>;
 
 export type SearchTheme = 'LIGHT' | 'DARK';
